+++ cdp/data_items/__init__.py
@@ -8,10 +8,16 @@ except:
 from os.path import dirname, basename
 import importlib
 
-module_files = glob.glob(dirname(__file__)+"/*.py")
+# Include .so's such that Cythonized cdp-py can still be properly imported
+module_files = glob.glob(dirname(__file__)+"/*.py") + glob.glob(dirname(__file__)+"/*.so")
 for module_file in module_files:
-    if not module_file.endswith('__init__.py'):
-        module_name = basename(module_file)[:-3]
+    if not '__init__' in module_file:
+        module_name = ''
+        if module_file.endswith('.py'):
+            module_name = basename(module_file)[:-3]
+        else:
+            # Cython outputs to module_name.cpython-python_version-platform.so
+            module_name = basename(module_file).split(".cpython")[0]
 
         # Import the module
         module = importlib.import_module(__package__+'.'+module_name)